create Awards refactoring
