+++ src/middleware/awardsAuthMiddleware.js
@@ -1,0 +1,7 @@
+
+export const awardsAuthMiddleware=(req,res,next)=>{
+    if (!req.headers.authorization) {
+        return sendErrorResponse(res,"User not authorized")
+    }
+    next()
+}
