--- src/modules/Awards/responseMessages.js
@@ -1,6 +1,0 @@
-export const successMessage={
-   createAwardSucess: "User Created Award added Successfully.",
-   editAwardsSucess: "Selected Award Modified Successfully.",
-   fetchAwardsSucess: "Awards Fetched Successfully.",
-   deleteAwardsSucess:"Awards Deleted Successfully."
-}
